+++ download-video-gui.sh
@@ -3,8 +3,8 @@
 # SPDX-License-Identifier: MIT
 # ============================================================================
 # Name        : download-video-gui.sh
-# Version     : 2.1.25
-# Date        : 2026-08-20
+# Version     : 2.1.26
+# Date        : 2026-08-21
 # Description : Zenity GUI for MKV video, authenticated YouTube HLS, or audio.
 # ============================================================================