+++ install-gui.sh
@@ -3,8 +3,8 @@
 # SPDX-License-Identifier: MIT
 # ============================================================================
 # Name        : install-gui.sh
-# Version     : 2.1.25
-# Date        : 2026-08-20
+# Version     : 2.1.26
+# Date        : 2026-08-21
 # Description : Install or remove the per-user desktop launcher and icon.
 # ============================================================================